+++ generate-examples/src/index.ts
@@ -12,6 +12,7 @@ const ISSUER_URL = process.env.ISSUER_URL || 'https://smarthealth.cards/examples
 const exampleBundleUrls = [
   'http://build.fhir.org/ig/dvci/vaccine-credential-ig/branches/main/Bundle-Scenario1Bundle.json',
   'http://build.fhir.org/ig/dvci/vaccine-credential-ig/branches/main/Bundle-Scenario2Bundle.json',
+  'https://www.hl7.org/fhir/diagnosticreport-example-ghp.json'
 ];
 
 interface Bundle {
@@ -82,11 +83,15 @@ async function trimBundleForHealthCard(bundleIn: Bundle) {
   bundle.entry.forEach((e) => {
     e.fullUrl = resourceUrlMap[e.fullUrl.split('/').slice(-2).join('/')];
     function clean(r: any, path: string[] = ['Resource']) {
+
       if (r.resourceType === 'Patient') {
+        // TODO remove these `delete`s once sample bundles are aligned
+        // with the "name + DOB" profiling guidance
         delete r.telecom;
         delete r.communication;
         delete r.address;
       }
+
       if (path.length === 1) {
         delete r.id;
         delete r.meta;
@@ -98,10 +103,7 @@ async function trimBundleForHealthCard(bundleIn: Bundle) {
       if (r.coding) {
         delete r.text;
       }
-      // Address bug in hosted examples
-      if (r.system === 'https://phinvads.cdc.gov/vads/ViewCodeSystem.action?id=2.16.840.1.113883.12.292') {
-        r.system = 'http://hl7.org/fhir/sid/cvx';
-      }
+
       if (r.system && r.code) {
         delete r.display;
       }
@@ -117,7 +119,7 @@ async function trimBundleForHealthCard(bundleIn: Bundle) {
   return bundle;
 }
 
-function createHealthCardJwsPayload(fhirBundle: Bundle): Record<string, unknown> {
+function createHealthCardJwsPayload(fhirBundle: Bundle, types: string[]): Record<string, unknown> {
   return {
     iss: ISSUER_URL + _issuerUrlSuffix,
     iat: new Date().getTime() / 1000, // TODO: add not yet valid
@@ -125,9 +127,10 @@ function createHealthCardJwsPayload(fhirBundle: Bundle): Record<string, unknown>
       '@context': ['https://www.w3.org/2018/credentials/v1'],
       type: [
         'VerifiableCredential',
-        _healthCardUri,
         'https://smarthealth.cards#immunization',
         'https://smarthealth.cards#covid19',
+        _healthCardUri,
+        ...types
       ],
       credentialSubject: {
         fhirVersion: '4.0.1', // TODO: which fhirVersion is ok?
@@ -135,6 +138,21 @@ function createHealthCardJwsPayload(fhirBundle: Bundle): Record<string, unknown>
       },
     },
   };
+}
+
+const MAX_SINGLE_JWS_SIZE = 1195;
+const MAX_CHUNK_SIZE = 1191;
+const splitJwsIntoChunks = (jws: string): string[] => {
+  if (jws.length <= MAX_SINGLE_JWS_SIZE) {
+    return [jws];
+  }
+
+  let chunks = [];
+  for (let i = 0; i < jws.length / MAX_CHUNK_SIZE; i++) {
+    chunks.push(jws.slice(i * MAX_CHUNK_SIZE, (i + 1) * MAX_CHUNK_SIZE))
+  }
+
+  return chunks;
 }
 
 async function createHealthCardFile(jwsPayload: Record<string, unknown>): Promise<Record<string, any>> {
@@ -146,8 +164,8 @@ async function createHealthCardFile(jwsPayload: Record<string, unknown>): Promis
 }
 
 const SMALLEST_B64_CHAR_CODE = 45; // "-".charCodeAt(0) === 45
-const toNumericQr = (jws: string): QRCodeSegment[] => [
-  { data: _qrHeader, mode: 'byte' },
+const toNumericQr = (jws: string, chunkIndex: number, totalChunks: number): QRCodeSegment[] => [
+  { data: _qrHeader + ((totalChunks > 1) ? `${chunkIndex + 1}/${totalChunks}/` : ``), mode: 'byte' },
   {
     data: jws
       .split('')
@@ -158,30 +176,38 @@ const toNumericQr = (jws: string): QRCodeSegment[] => [
   },
 ];
 
-async function processExampleBundle(exampleBundleUrl: string) {
+async function processExampleBundle(exampleBundleUrl: string): Promise<{ fhirBundle: Bundle; payload: Record<string, unknown>; file: Record<string, any>; qrNumeric: string[]; qrSvgFiles: string[]; }> {
+  let types = exampleBundleUrl.match("vaccine") ? [
+    'https://smarthealth.cards#immunization',
+    'https://smarthealth.cards#covid19',
+  ] : [];
+
   const exampleBundleRetrieved = (await got(exampleBundleUrl).json()) as Bundle;
   if (_longFhirBundle) {
     exampleBundleRetrieved.entry.push(_longFhirBundle);
   }
   const exampleBundleTrimmedForHealthCard = await trimBundleForHealthCard(exampleBundleRetrieved);
-  const exampleJwsPayload = createHealthCardJwsPayload(exampleBundleTrimmedForHealthCard);
+  const exampleJwsPayload = createHealthCardJwsPayload(exampleBundleTrimmedForHealthCard, types);
   const exampleBundleHealthCardFile = await createHealthCardFile(exampleJwsPayload);
-  const qrSegments = toNumericQr(exampleBundleHealthCardFile.verifiableCredential[0]);
-  const exampleBundleHealthCardNumericQr = qrSegments.map(({ data }) => data).join('');
+  const jws = exampleBundleHealthCardFile.verifiableCredential[0] as string;
+  const jwsChunks = splitJwsIntoChunks(jws);
+  const qrSet = jwsChunks.map((c, i, chunks) => toNumericQr(c, i, chunks.length));
+  const exampleBundleHealthCardNumericQr = qrSet.map(qr => qr.map(({ data }) => data).join(''));
 
-  const exampleQrCode: string = await new Promise((resolve, reject) =>
-    QrCode.toString(qrSegments, { type: 'svg' }, function (err: any, result: string) {
-      if (err) return reject(err);
-      resolve(result as string);
-    }),
-  );
+  const exampleQrCodes: string[] = await Promise.all(
+    qrSet.map((qrSegments): Promise<string> => new Promise((resolve, reject) =>
+      QrCode.toString(qrSegments, { type: 'svg', errorCorrectionLevel: 'low' }, function (err: any, result: string) {
+        if (err) return reject(err);
+        resolve(result as string);
+      })
+    )));
 
   return {
     fhirBundle: exampleBundleTrimmedForHealthCard,
     payload: exampleJwsPayload,
     file: exampleBundleHealthCardFile,
     qrNumeric: exampleBundleHealthCardNumericQr,
-    qrSvg: exampleQrCode,
+    qrSvgFiles: exampleQrCodes,
   };
 }
 
@@ -200,25 +226,32 @@ async function generate(options: { outdir: string, testcase:string }) {
     const fileC = `${outputPrefix}c-jws-payload-minified${ouputSuffix}.json`;
     const fileD = `${outputPrefix}d-jws${ouputSuffix}.txt`;
     const fileE = `${outputPrefix}e-file${ouputSuffix}.smart-health-card`;
-    const fileF = `${outputPrefix}f-qr-code-numeric${ouputSuffix}.txt`;
-    const fileG = `${outputPrefix}g-qr-code${ouputSuffix}.svg`;
+
+    const fileF = example.qrNumeric.map((qr, i) => `${outputPrefix}f-qr-code-numeric-value-${i}${ouputSuffix}.txt`);
+    const fileG = example.qrSvgFiles.map((qr, i) => `${outputPrefix}g-qr-code-${i}${ouputSuffix}.svg`);
 
     fs.writeFileSync(`${options.outdir}/${fileA}`, JSON.stringify(example.fhirBundle, null, 2));
     fs.writeFileSync(`${options.outdir}/${fileB}`, JSON.stringify(example.payload, null, 2));
     fs.writeFileSync(`${options.outdir}/${fileC}`, JSON.stringify(example.payload));
     fs.writeFileSync(`${options.outdir}/${fileD}`, example.file.verifiableCredential[0]);
     fs.writeFileSync(`${options.outdir}/${fileE}`, JSON.stringify(example.file, null, 2));
-    fs.writeFileSync(`${options.outdir}/${fileF}`, example.qrNumeric);
-    fs.writeFileSync(`${options.outdir}/${fileG}`, example.qrSvg);
+    example.qrNumeric.forEach((qr, i) => {
+      fs.writeFileSync(`${options.outdir}/${fileF[i]}`, qr);
+    });
+
+    example.qrSvgFiles.forEach((qr, i) => {
+      fs.writeFileSync(`${options.outdir}/${fileG[i]}`, qr);
+    });
 
     const exampleEntry: string[] = [];
+
     exampleEntry.push(fileA);
     exampleEntry.push(fileB);
     exampleEntry.push(fileC);
     exampleEntry.push(fileD);
     exampleEntry.push(fileE);
-    exampleEntry.push(fileF);
-    exampleEntry.push(fileG);
+    fileF.forEach(f => exampleEntry.push(f))
+    fileG.forEach(f => exampleEntry.push(f))
     exampleIndex[i] = exampleEntry;
   });
 
@@ -226,7 +259,7 @@ async function generate(options: { outdir: string, testcase:string }) {
   fs.writeFileSync(
     `${options.outdir}/index.md`,
     '# Example Resources \n' +
-      exampleIndex.map((e, i) => `## Example ${i}\n\n` + e.map((f) => `* [${f}](./${f})`).join('\n')).join('\n\n'),
+    exampleIndex.map((e, i) => `## Example ${i}\n\n` + e.map((f) => `* [${f}](./${f})`).join('\n')).join('\n\n'),
   );
 }